Adjusting payment id attribute to integer to match with Order id attribute

+++ src/main/java/dev/marcelomachado/crudspringjpapostgreesql/entities/Payment.java
@@ -12,7 +12,7 @@ public class Payment implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
 
     private Instant moment;
@@ -24,17 +24,17 @@ public class Payment implements Serializable {
     public Payment() {
     }
 
-    public Payment(Long id, Instant moment, Order order) {
+    public Payment(Integer id, Instant moment, Order order) {
         this.id = id;
         this.moment = moment;
         this.order = order;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
